rename: 폴더구조 변경 (shared 폴더 추가)
